% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_whatsapp}
\alias{parse_whatsapp}
\title{Parse a WhatsApp chat history}
\usage{
parse_whatsapp(path = "raw_data/_chat_example")
}
\arguments{
\item{path}{A path to an exported WhatsApp chat.}
}
\description{
Parse a WhatsApp chat history
}
