% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{term_table}
\alias{term_table}
\title{Create a tidy term frequency table}
\usage{
term_table(text, sort = term, ...)
}
\arguments{
\item{text}{A character vector or an object inheriting from \code{TextDocument}.}

\item{sort}{Unquoted name of column to sort by. Either term or freq.}

\item{...}{Further arguments passed to \code{\link[tm:termFreq]{tm::termFreq()}}.}
}
\description{
Create a tidy term frequency table
}
